Moved the service logic out of the oembed controller
refs #12537
